Extract error information from JSON.
When OSRM returns an error, it (often?) includes an error code/message
in the returned JSON object. This change extracts this code/message and
returns it to the caller of libosrmc. This makes it easier to debug
failures, but also makes it possible to react to errors such as
`NoRoute`.